Refactor shock capturing module into clean, modular structure
