+++ CHECKLIST.md
@@ -1,6 +1,6 @@
 # Development Checklist
 
-## Version 2.37.0 - Production Quality
+## Version 2.38.0 - Production Quality
 
 **Status: Active Development - Domain Modularization Complete**
 **Grade: A+ (95%)**
@@ -9,17 +9,17 @@
 
 ## Current Sprint Results
 
-### ✅ Completed This Sprint (Sprint 38 - Critical Code Review)
-- [x] **REFACTORED**: focused_transducer.rs (786 lines) into 4 domain modules
-- [x] **REMOVED**: Redundant kuznetsov_wave.rs (documentation-only file)
-- [x] **FIXED**: Adjective-based naming violations (acoustic_wave_kernel_optimized)
-- [x] **REPLACED**: Magic numbers with named constants (WATER_SOUND_SPEED)
-- [x] **ENFORCED**: SSOT principle throughout codebase
-- [x] **VALIDATED**: Physics implementations against literature
-- [x] **APPLIED**: Architecture principles (SOLID, CUPID, GRASP)
+### ✅ Completed This Sprint (Sprint 39 - Deep Code Audit)
+- [x] **REFACTORED**: shock_capturing.rs (782 lines) into 3 clean modules
+- [x] **FIXED**: ALL incomplete GPU kernel implementations  
+- [x] **REPLACED**: ALL placeholders with proper implementations
+- [x] **IMPLEMENTED**: Complete CUDA Level 3 register-blocked kernel
+- [x] **ADDED**: CPU fallback for GPU kernels ensuring correctness
+- [x] **REPLACED**: ALL magic numbers with named constants
+- [x] **VALIDATED**: All numerical methods against literature
 
 ### 🔄 In Progress
-- [ ] Refactoring 42 modules >500 lines (reduced from 43)
+- [ ] Refactoring 41 modules >500 lines (reduced from 42)
 - [ ] Reducing 453 warnings (increased due to new modules)
 
 ### 📋 Backlog
@@ -36,7 +36,7 @@
 | **Build Errors** | 0 | 0 | ✅ |
 | **Test Failures** | 0 | 0 | ✅ |
 | **Warnings** | 442 | <50 | ↓ |
-| **Modules >500 lines** | 42 | 0 | ↓ |
+| **Modules >500 lines** | 41 | 0 | ↓ |
 | **Modules >800 lines** | 1 | 0 | ↓ |
 | **Test Coverage** | 100% | 100% | ✅ |
 | **Examples Working** | 7/7 | 7/7 | ✅ |
